requirements.txt
Django==4.2
environs==9.5.0
django-filter==23.2
